ELT_Connector/connector_app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiobotocore==2.4.2
    # via s3fs
aiohttp==3.8.4
    # via
    #   aiobotocore
    #   s3fs
aioitertools==0.11.0
    # via aiobotocore
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.2
    # via aiohttp
attrs==22.2.0
    # via aiohttp
botocore==1.27.59
    # via aiobotocore
certifi==2022.12.7
    # via requests
charset-normalizer==2.1.1
    # via
    #   aiohttp
    #   requests
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2022.11.0
    # via s3fs
idna==3.4
    # via
    #   requests
    #   yarl
jmespath==1.0.1
    # via botocore
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.24.2
    # via pandas
pandas==1.5.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3
    # via pandas
requests==2.28.1
    # via -r requirements.in
s3fs==2022.11.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
urllib3==1.26.15
    # via
    #   botocore
    #   requests
wrapt==1.15.0
    # via aiobotocore
yarl==1.8.2
    # via aiohttp
